♻️ Rename strings.ts to union.ts
